Fix when we do the field filtering to be after we get the names and handle appearance
A previous fix for the search skills broke the processor in prod that sends out phase change emails :(

If the query includes field parameters that *don’t* include `handleLower` or `userId`, when we query for names and handle appearance traits, we end up throwing an exception because there’s nothing to search on.

The fix was to move the field filtering after we’ve done the additional lookups, to ensure that we always have the expected query parameters for the additional searching.